+++ preproc/runoff/match_rivers.m
@@ -1,102 +1,71 @@
 function river_info = match_rivers(lon_elems, lat_elems, ind_elems)
 % Match the source elements with rivers
 %
-%% Syntax
-% river_list = match_rivers(lon_elems, lat_elems)
+%% Syntax:
+%   river_info = match_rivers(lon_elems, lat_elems, ind_elems)
 %
-%% Description
-% river_list = match_rivers(lon_elems, lat_elems) matches the source
-% elements with rivers.
+%% Description:
+%   Matches source elements (specified by longitude and latitude) with
+%   pre-defined estuarine zones representing major rivers.
 %
-%% Examples
-% river_list = match_rivers(lon_elems, lat_elems)
+%% Input Arguments:
+%   lon_elems --- vector of longitudes
+%   lat_elems --- vector of latitudes
+%   ind_elems --- indices of the source elements
 %
-%% Input Arguments
-% lon_elems --- a vector containing the longitude of source elements.
-% lat_elems --- a vector containing the latitude of source elements.
-%
-%% Output Arguments
-% river_list --- corresponding rivers for all given elements. There can be
-% dulplicate rivers, meaning that more than one outlets for the same river.
-%
-%% Notes
-% User can easily add rivers in the code, and the only variable you need to
-% provide is the 'estuary_zone', namely the region covering the river
-% estuary, e.g. estuary_zone = [119.5 122 31.5 32.5] denotes the Changjiang
-% River. 
+%% Output Arguments:
+%   river_info.River  --- river names of matched elements
+%   river_info.Lon    --- longitudes of matched elements
+%   river_info.Lat    --- latitudes of matched elements
+%   river_info.Elements --- indices of matched elements
 %
 %% Author Info
-% Created by Wenfan Wu, Ocean Univ. of China in 2023.
-% Last Updated on 2023-11-24.
-% Email: wenfanwu@stu.ouc.edu.cn
+% Created by Wenfan Wu, Virginia Institute of Marine Science in 2023.
+% Last Updated on 6 May 2025.
+% Email: wwu@vims.edu
 %
-% See also: def_schism_source and add_river_runoff
+% See also: add_river_runoff
 
-%% Parse inputs
-river_list = cell(size(lon_elems));
-river_list(:) = {'unknown'};
+%% Define estuarine zones for rivers (CAN BE CHANGED)
+river_definitions = {'Huanghe',       [118.8 119.6 37.5 38.1];
+    'Huaihe',        [120.2 120.45 34 34.2];
+    'Haihe',         [117.6 117.88 38.9 39.1];
+    'Luanhe',        [119.2 119.4 39.2 39.55];
+    'Liaohe',        [121.6 122 40.73 41.2];
+    'Changjiang',    [119.5 122 31.5 32.5];
+    'Qiantangjiang', [120 122 30 31];
+    'Yalujiang',     [124.1693 124.4999 39.7008 40.0907];
+};
 
-estuary_zone = [118.8 119.6 37.5 38.1];
-ind_river  = is_river(lon_elems, lat_elems, estuary_zone);
-river_list(ind_river) = {'Huanghe'};
+%% Begin to match rivers
+% Initialize river info
+river_info(numel(lon_elems), 1) = struct('River', [], 'Lon', [], 'Lat', [], 'Elements', []);
 
-estuary_zone = [117.6 117.88 38.9 39.1];
-ind_river  = is_river(lon_elems, lat_elems, estuary_zone);
-river_list(ind_river) = {'Haihe'};
-
-estuary_zone = [119.2 119.4 39.2 39.55];
-ind_river  = is_river(lon_elems, lat_elems, estuary_zone);
-river_list(ind_river) = {'Luanhe'};
-
-estuary_zone = [121.6 122 40.73 41.2];
-ind_river  = is_river(lon_elems, lat_elems, estuary_zone);
-river_list(ind_river) = {'Liaohe'};
-
-estuary_zone = [119.5 122 31.5 32.5];
-ind_river  = is_river(lon_elems, lat_elems, estuary_zone);
-river_list(ind_river) = {'Changjiang'};
-
-estuary_zone = [120 122 30 31];
-ind_river  = is_river(lon_elems, lat_elems, estuary_zone);
-river_list(ind_river) = {'Qiantangjiang'};
-
-% remove unvalid source elements
-ind_unvalid = contains(river_list, 'unknown');
-
-if nargin>=3
-    river_info.river_elems = ind_elems;
+% Match each river
+for ii = 1:size(river_definitions, 1)
+    river_name = river_definitions{ii,1};
+    river_zone = river_definitions{ii,2};
+    ind = is_river(lon_elems, lat_elems, river_zone);
+    for jj = 1:numel(ind)
+        river_info(ind(jj)).River = river_name;
+        river_info(ind(jj)).Lon = lon_elems(ind(jj));
+        river_info(ind(jj)).Lat = lat_elems(ind(jj));
+        river_info(ind(jj)).Elements = ind_elems(ind(jj));       
+    end
 end
-river_info.lon_list = lon_elems(~ind_unvalid);
-river_info.lat_list = lat_elems(~ind_unvalid);
-river_info.river_list = river_list(~ind_unvalid);
 
-if numel(find(ind_unvalid))>0
-    warning([num2str(numel(find(ind_unvalid))), ' unvalid source elements have been removed'])
+% Display summary
+ind_empty = arrayfun(@(x) isempty(x.River), river_info);
+if any(ind_empty)
+    warning('%d unvalid source elements have been removed', sum(ind_unvalid));
 else
-    disp('all selected elements are matched!')
+    disp('All selected elements are matched!');
+end
+river_info = river_info(~ind_empty);  % Remove empty rows.
+
 end
 
+function ind_river = is_river(lon, lat, zone)
+% Check if coordinates fall within the estuarine zone
+ind_river = find(lon > zone(1) & lon < zone(2) & lat > zone(3) & lat < zone(4));
 end
-
-function ind_river  = is_river(lon_elems, lat_elems, estuary_zone)
-% Determine whether the given point is inside a certain region
-
-ind_river = lon_elems > estuary_zone(1) & lon_elems < estuary_zone(2) & lat_elems > estuary_zone(3) & lat_elems < estuary_zone(4);
-end
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-